+++ hoot-core-test/src/test/cpp/hoot/core/visitors/RandomElementRemoverTest.cpp
