+++ hoot-core/src/main/cpp/hoot/core/ops/RandomNodeDuplicator.h
@@ -44,8 +44,6 @@ namespace hoot
  * Duplicates a POI with some random error.  This is not part of the original Perty paper.
  *
  * This should really change the node after it has been duplicated.
- *
- * @todo This duplicates nodes, not POIs, so the name should be corrected.
  */
 class PertyDuplicatePoiOp : public OsmMapOperation, public Configurable, public RngConsumer,
   public OperationStatusInfo
