+++ hoot-core/src/main/cpp/hoot/core/visitors/RandomElementRemover.cpp
