+++ hoot-core/src/main/cpp/hoot/core/visitors/RandomElementRenamer.cpp
