+++ hoot-core/src/main/cpp/hoot/core/visitors/RandomTagModifier.h
@@ -24,62 +24,60 @@
  *
  * @copyright Copyright (C) 2015, 2017, 2018, 2019 DigitalGlobe (http://www.digitalglobe.com/)
  */
-#ifndef PERTYREMOVERANDOMELEMENTVISITOR_H
-#define PERTYREMOVERANDOMELEMENTVISITOR_H
+#ifndef PERTY_REMOVE_TAG_VISITOR_H
+#define PERTY_REMOVE_TAG_VISITOR_H
 
 // hoot
-#include <hoot/core/info/OperationStatusInfo.h>
+#include <hoot/core/elements/ElementVisitor.h>
 #include <hoot/core/util/Configurable.h>
-#include <hoot/core/visitors/ElementOsmMapVisitor.h>
 #include <hoot/core/algorithms/perty/RngConsumer.h>
 
 namespace hoot
 {
 
 /**
- * Randomly removes elements from a map
- *
- * @todo rename to RandomElementRemover
+ * A simple random tag remover
  */
-class PertyRemoveRandomElementVisitor : public ElementOsmMapVisitor, public RngConsumer,
-    public Configurable, public OperationStatusInfo
+class PertyRemoveTagVisitor : public ElementVisitor, public RngConsumer, public Configurable
 {
 public:
 
-  static std::string className() { return "hoot::PertyRemoveRandomElementVisitor"; }
+  static std::string className() { return "hoot::PertyRemoveTagVisitor"; }
 
-  /**
-   * @arg p Probability that any given feature will be removed.
-   */
-  PertyRemoveRandomElementVisitor();
+  PertyRemoveTagVisitor();
+
+  QString permuteName(const QString& s);
 
   virtual void setConfiguration(const Settings& conf);
 
   /**
-   * Set the probability that a feature will be removed.
+   * Set the probability that a tag will be removed.
    */
   void setProbability(double p) { _p = p; }
 
   virtual void setRng(boost::minstd_rand& rng) { _rng = &rng; }
 
-  virtual void visit(const ConstElementPtr& e);
-  virtual void visit(const std::shared_ptr<Element>& /*e*/) override {}
+  virtual void visit(const std::shared_ptr<Element>& e) override;
 
-  virtual QString getDescription() const { return "Randomly removes elements from a map"; }
+  void setExemptTagKeys(const QStringList& keys) { _exemptTagKeys = keys; }
 
-  virtual QString getInitStatusMessage() const
-  { return "Removing random elements..."; }
+  void setReplacementTagKeys(const QStringList& keys) { _replacementTagKeys = keys; }
 
-  virtual QString getCompletedStatusMessage() const
-  { return "Randomly removed " + QString::number(_numAffected) + " elements"; }
+  void setReplacementTagValues(const QStringList& values) { _replacementTagValues = values; }
+
+  virtual QString getDescription() const { return "Randomly removes feature tags"; }
 
 private:
 
+  double _p;
   boost::minstd_rand* _rng;
   std::shared_ptr<boost::minstd_rand> _localRng;
-  double _p;
+
+  QStringList _exemptTagKeys;
+  QStringList _replacementTagKeys;
+  QStringList _replacementTagValues;
 };
 
 }
 
-#endif // PERTYREMOVERANDOMELEMENTVISITOR_H
+#endif // PERTY_REMOVE_TAG_VISITOR_H
