+++ hoot-core/src/main/cpp/hoot/core/visitors/RandomWayGeneralizer.cpp
